requirements.txt
# OpenSSL Conan Development Environment
# Essential dependencies for cross-platform Conan package development

# Core Conan dependencies
conan>=2.0.0
PyYAML>=6.0
requests>=2.28.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality (optional)
flake8>=6.0.0
black>=22.0.0

# Security scanning (optional)
safety>=2.3.0
bandit>=1.7.0

# Utilities
click>=8.1.0
rich>=13.0.0

# Development tools (optional)
pre-commit>=3.0.0
mypy>=1.0.0

# CI/CD automation
GitPython>=3.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# MCP Server Dependencies
mcp>=1.0.0
fastmcp>=0.9.0
httpx>=0.25.0
tenacity>=8.2.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0

# CI/CD and Automation
semgrep>=1.0.0
codecov>=2.1.0

# GitHub Integration
PyGithub>=2.1.1
gitpython>=3.1.40

# OpenSSL CI/CD Requirements
# Used for caching optimization in GitHub Actions
conan==2.0.17
# trivy>=0.50.0  # Commented out - not available via pip

# Optional advanced features (uncomment as needed)
# pytest-xdist>=3.0.0  # Parallel testing
# pytest-html>=3.0.0   # HTML test reports
# pylint>=2.15.0       # Advanced linting
# isort>=5.12.0        # Import sorting
# psutil>=5.9.0        # System monitoring
# prometheus-client>=0.15.0  # Metrics collection
# sphinx>=5.0.0        # Documentation generation
# cyclonedx-bom>=3.0.0 # SBOM generation
